requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
fastapi-mail==1.4.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
motor==3.7.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
websockets==13.1

python-multipart==0.0.6
